+++ packages/camera/camera_avfoundation/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.9.17+8
+
+* Introduces new protocols `FLTEventChannel` and `FLTCameraDeviceDiscovering`
+
+## 0.9.17+7
+
+* Introduces new protocols `FLTCaptureDeviceControlling` and `FLTDeviceOrientationProviding`
+
 ## 0.9.17+6
 
 * Updates minimum supported SDK version to Flutter 3.22/Dart 3.4.